Move folder into api_docs/files
