+++ apis/client/import_documentation/import.js
@@ -7,7 +7,7 @@ import { URI } from 'meteor/olragon:uri-js';
 import _ from 'lodash';
 import jsyaml from 'js-yaml';
 
-import DocumentationFiles from '/doc_files/collection';
+import DocumentationFiles from '/api_docs/files/collection';
 
 Template.importApiDocumentation.events({
   'change #apiDocumentationFile': function (event) {
